PageInfoBubbleView was refactored.
Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/3ed46b480719847f8745ddf43c2d39edeead6156

commit 3ed46b480719847f8745ddf43c2d39edeead6156
Author: Olesia Marukhno <olesiamarukhno@google.com>
Date:   Tue Mar 23 15:26:39 2021 +0000

    Refactor PageInfoBubbleView

    Removed references to web_contents in bubble and mostly removed
    references to profile. This is needed for future redesign, to keep new
    views independent from web_contents and profile and keep all logic in
    presenter and delegates.

    Bug: 1188101
